+++ project-types/CLILibrary/src/@i18n/index.ts
@@ -10,7 +10,8 @@
  * You may read the full license at https://gobstones.github.io/gobstones-guidelines/LICENSE.
  * *****************************************************************************
  */
-/* eslint-disable import/no-extraneous-dependencies */
+
+/* eslint-disable import/no-extraneous-dependencies, import/export*/
 /**
  * ----------------------------------------------------
  * @module i18n
@@ -27,17 +28,15 @@ import i18next from 'i18next';
 import i18nextCLILanguageDetector from 'i18next-cli-language-detector';
 import i18nextBackend from 'i18next-fs-backend';
 
-function currentPath() {
+const currentPath = (): string => {
     const currentModuleDir = __dirname;
     const translationsFolder = '@i18n';
     let nextDir = currentModuleDir;
     while (!fs.existsSync(path.join(nextDir, translationsFolder))) {
         nextDir = path.dirname(nextDir);
     }
-    // eslint-disable-next-line no-console
-    console.log(path.join(nextDir, translationsFolder));
     return path.join(nextDir, translationsFolder);
-}
+};
 
 void i18next
     .use(i18nextCLILanguageDetector)
@@ -45,7 +44,7 @@ void i18next
     .init({
         initImmediate: false,
         fallbackLng: 'en',
-        ns: ['application', 'cli'],
+        ns: ['application'],
         backend: {
             loadPath: currentPath() + '/{{lng}}/{{ns}}.json'
         }