+++ project-types/Common/typedoc.config.mjs
@@ -12,23 +12,15 @@ export default {
         rootDir: `${rootDir}/src`
     },
     out: `${rootDir}/docs`,
-    exclude: [`${rootDir}/node_modules/**/*`, `${rootDir}/**/*.test.ts`, `${rootDir}/src/index.ts`],
-    includeVersion: true,
-    categorizeByGroup: true,
+    exclude: [
+        `${rootDir}/node_modules/**/*`,
+        `${rootDir}/**/*.test.ts`,
+        `${rootDir}/src/@types/**/*`,
+        `${rootDir}/src/index.ts`
+    ],
+    plugin: ['@gobstones/typedoc-theme-gobstones'],
+    theme: 'gobstones',
     excludeExternals: true,
     excludeInternal: false,
-    excludePrivate: false,
-    hideGenerator: true,
-    disableSources: false,
-    githubPages: true,
-    plugin: ['@gobstones/typedoc-theme-gobstones'],
-    excludeTags: ['@override', '@virtual', '@satisfies', '@overload'],
-    visibilityFilters: {
-        '@internal': false,
-        protected: false,
-        private: false,
-        inherited: false
-    },
-    theme: 'gobstones',
-    mergeModulesMergeMode: 'module-category'
+    excludePrivate: false
 };